INT-4390: Phase II - dsl<->dsl.channel tangles
Move the channel specs to dsl.
