+++ spring-integration-core/src/main/java/org/springframework/integration/dsl/PublishSubscribeSpec.java
@@ -20,8 +20,6 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.concurrent.Executor;
 
-import org.springframework.integration.dsl.channel.PublishSubscribeChannelSpec;
-
 /**
  * @author Artem Bilan
  *
