--- spring-integration-core/src/main/java/org/springframework/integration/dsl/channel/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains MessageChannel Builders DSL.
- */
-package org.springframework.integration.dsl.channel;
