+++ spring-integration-file/src/test/java/org/springframework/integration/file/dsl/FileTests.java
@@ -64,9 +64,9 @@ import org.springframework.integration.config.EnableIntegration;
 import org.springframework.integration.dsl.IntegrationFlow;
 import org.springframework.integration.dsl.IntegrationFlowDefinition;
 import org.springframework.integration.dsl.IntegrationFlows;
+import org.springframework.integration.dsl.MessageChannels;
 import org.springframework.integration.dsl.Pollers;
 import org.springframework.integration.dsl.StandardIntegrationFlow;
-import org.springframework.integration.dsl.channel.MessageChannels;
 import org.springframework.integration.expression.FunctionExpression;
 import org.springframework.integration.file.DefaultDirectoryScanner;
 import org.springframework.integration.file.DefaultFileNameGenerator;
